requirements.txt
langgraph==0.2.16
langchain==0.3.0
langchain-ollama==0.2.0
requests==2.32.3
python-dotenv==1.0.1